Update project - Clean architecture - create package
