+++ app/src/main/java/com/example/moves/domain/Poster.kt
@@ -1,4 +1,4 @@
-package com.example.moves
+package com.example.moves.domain
 
 import java.io.Serializable
