+++ app/src/main/java/com/example/moves/domain/ReviewResponse.kt
@@ -1,5 +1,6 @@
-package com.example.moves
+package com.example.moves.domain
 
+import com.example.moves.domain.Review
 import com.google.gson.annotations.SerializedName
 
 class ReviewResponse(@SerializedName("docs") val reviewList: List<Review>)
