+++ app/src/main/java/com/example/moves/domain/TrailerResponse.kt
@@ -1,5 +1,6 @@
-package com.example.moves
+package com.example.moves.domain
 
+import com.example.moves.domain.Video
 import com.google.gson.annotations.SerializedName
 
 class TrailerResponse(@SerializedName("videos") val trailsList: Video) {
