+++ app/src/main/java/com/example/moves/domain/Video.kt
@@ -1,5 +1,6 @@
-package com.example.moves
+package com.example.moves.domain
 
+import com.example.moves.domain.Trailer
 import com.google.gson.annotations.SerializedName
 
 class Video(@SerializedName("trailers") val trailers: List<Trailer>)
