+++ app/src/main/java/com/example/moves/presentation/screens/main/MainViewModule.kt
@@ -1,4 +1,4 @@
-package com.example.moves
+package com.example.moves.presentation.screens.main
 
 import android.app.Application
 import android.util.Log
@@ -6,6 +6,7 @@ import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import com.example.moves.data.network.ApiFactory
+import com.example.moves.domain.Movie
 import io.reactivex.rxjava3.android.schedulers.AndroidSchedulers
 import io.reactivex.rxjava3.disposables.CompositeDisposable
 import io.reactivex.rxjava3.functions.Action
