+++ app/src/main/res/layout/activity_favorite_movies.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
+    tools:context=".presentation.screens.main.MainActivity">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/favoriteMovieRecyclerView"
